+++ feeluown/ui.py
@@ -15,6 +15,7 @@ from PyQt5.QtWidgets import (
     QVBoxLayout,
 )
 
+from fuocore.models import Media
 from fuocore.player import PlaybackMode, State
 from fuocore.utils import parse_ms
 
@@ -490,7 +491,7 @@ class Ui:
 
     def _play_mv(self):
         song = self._app.player.current_song
-        url = song.mv.media.url_ahap
+        url = song.mv.media.select_url(Media.Q.sq)
         self._app.player.play(url)
         self.show_video_widget()